requirements.txt
playwright>=1.41.0
Pillow>=10.0.0
